+++ bostonbusmap-library/src/main/java/com/schneeloch/bostonbusmap_library/data/IntersectionLocation.java
@@ -108,7 +108,7 @@ public class IntersectionLocation implements Location {
 	}
 
 	@Override
-	public float distanceFromInMiles(double centerLatAsRadians,
+	public double distanceFromInMiles(double centerLatAsRadians,
 			double centerLonAsRadians) {
 		return Geometry.computeDistanceInMiles(latitude, longitude, centerLatAsRadians, centerLonAsRadians);
 	}
